Extract payment queue classes to a separate package
