--- src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/servicebus/IPaymentsPublisher.java
@@ -1,8 +1,0 @@
-package uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus;
-
-import uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.model.PaymentsData;
-
-public interface IPaymentsPublisher {
-
-    void publishPayments(PaymentsData paymentsData);
-}
